fix(dfn): restore old dfn2toml logic
